ApiTestFrame/requirements.txt
allure_python_commons==2.13.2
filelock==3.13.1
jsonpath==0.82
PyMySQL==1.0.3
pytest==7.4.0
PyYAML==6.0
requests==2.31.0
